make member set a hashmap
